requirements.txt
pywin32==306
Pillow==10.0.0
psutil==5.9.5 